Consilidate SID tests in single file in tests/unit
This commit fixes the jenkins unit test pipeline by consolidating
some newer sid tests into the tests/unit hierarchy.

(cherry picked from commit 5c42e16a4176ee64c9393246abc57b09584181c5)
